[TASK] Resolve `Model/AbstractTimespan` into `Model/Event`
As the `AbstractTimespan` class now has only one subclass,
there is no need for it to exist anymore.

Fixes #4734
