+++ Tests/LegacyFunctional/Model/EventDateTest.php
@@ -13,7 +13,6 @@ use TYPO3\TestingFramework\Core\Functional\FunctionalTestCase;
 /**
  * This test case holds all tests specific to event dates.
  *
- * @covers \OliverKlee\Seminars\Model\AbstractTimeSpan
  * @covers \OliverKlee\Seminars\Model\Event
  */
 final class EventDateTest extends FunctionalTestCase
